format add softdeletable to interface

+++ Handler/AbstractHandler.php
@@ -32,13 +32,29 @@ abstract class AbstractHandler implements HandlerInterface
     protected $softdeletable = true;
 
     /**
+     * @param bool $flag
+     * @return AbstractHandler
+     */
+    public function setSoftDeletable($flag){
+        $this->softdeletable = $flag;
+        return $this;
+    }
+
+    /**
+     * @return bool
+     */
+    public function getSoftdeletable()
+    {
+        return $this->softdeletable;
+    }
+
+    /**
      * @param ImporterInterface $importer
      * @return AbstractHandler
      */
     public function setImporter(ImporterInterface $importer)
     {
         $this->importer = $importer;
-
         return $this;
     }
 
@@ -49,7 +65,6 @@ abstract class AbstractHandler implements HandlerInterface
     public function setEntityManager(EntityManager $entityManager)
     {
         $this->entityManager = $entityManager;
-
         return $this;
     }
 
@@ -62,23 +77,28 @@ abstract class AbstractHandler implements HandlerInterface
      */
     protected function import($data, $className, $flush, OutputInterface $output = null)
     {
-        if (!$output) {
+        if(!$output){
             $output = new NullOutput();
         }
 
         $em = $this->entityManager;
-        if ($this->softdeletable) {
+
+        if($this->softdeletable){
             $em->getFilters()->disable('softdeleteable');
+            $output->writeln('<comment>---Disable softdeleteable---</comment>');
         }
+
         $output->writeln('<comment>---Start Import---</comment>');
 
         $importer = $this->importer;
         $resultBag = $importer->getResultBag();
 
         $importer->process($data, $className);
+
         $this->writeResultBagToOutput($output, $resultBag);
 
-        if (true === $flush) {
+        if(true === $flush){
+
             $reflection = new \ReflectionClass($className);
             if ($this->softdeletable && (!$reflection->hasProperty('deletedAt') || !$reflection->hasMethod('setDeletedAt') || !$reflection->hasMethod('getDeletedAt'))) {
                 throw new MethodNotFoundException(sprintf('It seems that the entity "%s" is not softdeleteable', $className));
@@ -86,13 +106,13 @@ abstract class AbstractHandler implements HandlerInterface
 
             $this->removeUnneededEntities($output, $resultBag);
 
-            if ($resultBag->hasChanges()) {
+            if ($resultBag->hasChanges()){
+
                 $entities = array_merge($resultBag->getNew(), $resultBag->getChanged());
-                foreach ($entities as $entity) {
-                    if ($this->softdeletable) {
+                foreach($entities as $entity){
+                    if($this->softdeletable){
                         $entity->setDeletedAt(null);
                     }
-
                     $em->persist($entity);
                 }
 
@@ -101,7 +121,8 @@ abstract class AbstractHandler implements HandlerInterface
             }
 
         }
-        if ($this->softdeletable) {
+
+        if($this->softdeletable){
             $em->getFilters()->enable('softdeleteable');
         }
     }
@@ -111,10 +132,10 @@ abstract class AbstractHandler implements HandlerInterface
         $em = $this->entityManager;
 
         $output->writeln('<error>Removing ' . $resultBag->countRemoving() . ' Entries</error>');
-        foreach ($resultBag->getRemoving() as $entity) {
-            //             if($entity->getDeletedAt() === null){
-            $em->remove($entity);
-            //             }
+        foreach ($resultBag->getRemoving() as $entity){
+            if(!$this->softdeletable || null === $entity->getDeletedAt()){
+                $em->remove($entity);
+            }
         }
 
         $em->flush();
@@ -131,16 +152,4 @@ abstract class AbstractHandler implements HandlerInterface
         $output->writeln('<info>Skipped: ' . $resultBag->countSkipped() . ' Entries</info>');
         $output->writeln('<error>New: ' . $resultBag->countNew() . ' Entries</error>');
     }
-
-    public function getSoftdeletable()
-    {
-        return $this->softdeletable;
-    }
-
-    public function setSoftdeletable($softdeletable)
-    {
-        $this->softdeletable = $softdeletable;
-        return $this;
-    }
-
 }
